# add len check in getExtendedPathSpecs release-4.3.9