--- a/cli/linter/schema.json
+++ b/cli/linter/schema.json
@@ -959,6 +959,9 @@
     "suppress_redis_signal_reload": {
       "type": "boolean"
     },
+    "reload_interval": {
+      "type": "integer"
+    },
     "syslog_network_addr": {
       "type": "string"
     },