<!-- Provide a general summary of your changes in the Title above -->

## Description

Before 
```
pkg: github.com/TykTechnologies/tyk/gateway
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_getExtendedPathSpecs-16        2804581               860.0 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2651720               884.0 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2596935               938.3 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2541750               900.8 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2633457               926.0 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2711748               907.3 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2602863               938.3 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2557484               956.0 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2525793               944.2 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2752916               872.7 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/TykTechnologies/tyk/gateway  34.029s
```

After
```
pkg: github.com/TykTechnologies/tyk/gateway
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_getExtendedPathSpecs-16       28540504                80.53 ns/op            0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       29428118                77.38 ns/op            0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       28900099                77.02 ns/op            0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       29882416                76.86 ns/op            0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       31099144                76.46 ns/op            0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       30997551                76.47 ns/op            0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       28353508                75.93 ns/op            0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       31181818                76.33 ns/op            0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       30218952                75.84 ns/op            0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       30840214                76.68 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/TykTechnologies/tyk/gateway  24.702s
```
## Related Issue

<!-- This project only accepts pull requests related to open issues. -->
<!-- If suggesting a new feature or change, please discuss it in an issue first. -->
<!-- If fixing a bug, there should be an issue describing it with steps to reproduce. -->
<!-- OSS: Please link to the issue here. Tyk: please create/link the JIRA ticket. -->

## Motivation and Context

<!-- Why is this change required? What problem does it solve? -->

## How This Has Been Tested

<!-- Please describe in detail how you tested your changes -->
<!-- Include details of your testing environment, and the tests -->
<!-- you ran to see how your change affects other areas of the code, etc. -->
<!-- This information is helpful for reviewers and QA. -->

## Screenshots (if appropriate)

## Types of changes

<!-- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Refactoring or add test (improvements in base code or adds test coverage to functionality)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes that apply -->
<!-- If there are no documentation updates required, mark the item as checked. -->
<!-- Raise up any additional concerns not covered by the checklist. -->

- [ ] I ensured that the documentation is up to date
- [ ] I explained why this PR updates go.mod in detail with reasoning why it's required
- [ ] I would like a code coverage CI quality gate exception and have explained why