@Tyk-ITS 2023-08-28
**API tests result: failure** :no_entry_sign: 
Branch used: refs/pull/5471/merge
Commit:  
Triggered by: pull_request (@jeffy-mathew)
[Execution page](https://github.com/TykTechnologies/tyk/actions/runs/5997192495)