@Tyk-ITS 2023-08-28
**API tests result: failure** :no_entry_sign: 
Branch used: refs/pull/5471/merge
Commit: f0fafe7fdc6052ce44277225ee22778c27d0113c 
Triggered by: pull_request (@jeffy-mathew)
[Execution page](https://github.com/TykTechnologies/tyk/actions/runs/5997838380)