@Tyk-ITS 2023-08-28
**API tests result: failure** :no_entry_sign: 
Branch used: refs/pull/5471/merge
Commit: 935543d3cf54e9119a050c31240e6f226d0f0f9b 
Triggered by: pull_request (@jeffy-mathew)
[Execution page](https://github.com/TykTechnologies/tyk/actions/runs/5998977681)